… CHANGELOG.md
… docs/sources/configuration/_index.md