**What this PR does / why we need it**:
Experimental feature Optimize json parser in limited cases.
Implement a high-performance simplejson parser（labels_hints_fast_parse）.
```shell
Benchmark_JsonParserFastParse/json/no_labels_hints-12         	  272154	      4296 ns/op
Benchmark_JsonParserFastParse/json/labels_hints_fast_parse-12 	 4061995	       258.0 ns/op
```
for example:
```go
logql ：
{functionName="feedflow"}  | json   | durationMs > 2003 
transform to
Optimize logql ：
{ functionName="feedflow"} | simplejson | durationMs>2003 | json 
```

In this way, reduce the performance loss caused by full json key parsing

**Which issue(s) this PR fixes**:
Fixes #5405>

**Special notes for your reviewer**:
"linefmt" case logql is more complicated, and another PR will handle it.
This feature is turned off by default, I will turn on this switch in our loki cluster to continuously test and trace this feature

loki.yaml config
```yaml
querier:
  query_timeout: 30m
  engine:
    logql_optimize_enable: true
  query_ingesters_within: 2h
```

**online environment**
I have completed the test in the online environment
tracing
![image](https://user-images.githubusercontent.com/9583245/154459262-ba9a101d-d8b3-4884-8288-0142797037ad.png)
+ performance = 50s-39/50s = 22%.  I will continue profile this json issue.

with this pr 39s
![image](https://user-images.githubusercontent.com/9583245/154459489-7d1eb4dd-d12a-4276-b44b-c8fd6584e7e5.png)
without this pr 50s
![image](https://user-images.githubusercontent.com/9583245/154459616-6c3d3c4e-9714-4f60-8b75-f0723c4359cc.png)
**Checklist**
- [x] Have you completed the test in the online environment
- [x] Documentation added
- [x] Tests updated
- [x] Add an entry in the `CHANGELOG.md` about the changes.