@splitice 2022-02-17
Isnt the test a no-op?

`{log_type="service_metrics", module="api_server", operation="InvokeFunction", accountID="212068714932184585", serviceName="taojimu-fc-prod", functionName="feedflow"} | json | durationMs>2003 | json`

Or is this some new magic?