@liguozhong 2022-02-24
> I support this idea, I think this could be better specially using hints, but I disagree with the implementation I think this should be the default path not something experimental or a replacement of LogQL queries.
> 
> My suggestion is to look into https://github.com/buger/jsonparser or even simdjson and replace the current parser which does an allocation for each key at minimum.
> 
> But I don't think a custom parsing will do it, there's so many special case to handle like quoting, escaping, etc....

Thanks for your review。
 I was busy with the coding test of another PR a few days ago (https://github.com/grafana/loki/pull/5455). 

Now that the test has been completed, I will start to see how the json library you suggested is used ，and try to optimize this PR