@cyriltovena 2022-02-24
Appreciate the work, and I do think it has value since I created an issue. Just skeptic about the current direction.