@liguozhong 2022-03-04
> Appreciate the work, and I do think it has value since I created an issue. Just skeptic about the current direction.

https://github.com/grafana/loki/pull/5538, I submit another PR